fix unmount causing lost context
