+++ web/frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@viamrobotics/remote-control",
-  "version": "2.0.11",
+  "version": "2.0.12",
   "license": "Apache-2.0",
   "type": "module",
   "files": [
